Updated the Exception classes, also added a generic exception.
